Rename ftree package to tree
Avoids naming collision for `go build`
